requirements.txt
Jinja2~=2.11.2
python-dotenv~=0.15.0
requests~=2.25.1
portage~=3.0.14
typer~=0.3.2
setuptools~=49.2.1
pytest~=6.2.2
toolz~=0.11.1